gcp/terraform/function/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*